Added basic support for a postgresql driver
* Base model templates have been updated to escape quotes in the relation queries, since PG uses a double quote for denoting object names (rather than the MySQL back tick)
* Columns that are `auto_increment` or `sequence` are noted as being `serial`. When inserting a new row the column value is set to `DEFAULT` (unless it has been overridden post-`__construct`)
* A few sundry utility methods
* Removed the `LIMIT` on `UPDATE` statements since PG doesn't support this. Not sure if this has any further ramifications.
* Detection when adding duplicate columns/relations to tables in the generator.

Overall, there is an inconsistent use `schema` and `database`, since MySQL's language is not strictly the same as PostgreSQL. This could be addressed by moving to the more formal language used by PostgreSQL, and adapting the MySQL driver to fill in the blanks. Not sure if this makes sense or not, need further discussion about the intricacies of language between the two.