Only regenerate base models if they have changed
Adds a 'fingerprint' to the class docblock for each generated base
model. This fingerprint is a MD5 hash of the serialised table
schema as extracted by the DB driver.

The generator then calculates this hash and if its the same it
doesn't replace the file. This means running the model generator
is now idempotent. Previously files would always be touched with
an updated @generated timestamp, even if the schema hadn't
changed.